requirements.txt
playwright==1.41.2
requests==2.31.0
python-dotenv
